requirements.txt
argon2-cffi==19.2.0
celery==4.4.2
dictdiffer==0.8.1
Django==3.0.5
django-celery-beat==2.0.0
django-cors-headers==3.3.0
django-filter==2.2.0
django-imagekit==4.0.2
django-templated-email==2.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
gevent==20.5.0
gunicorn==20.0.4
Pillow==7.1.1
psycopg2-binary==2.8.5
sentry-sdk==0.14.4
requests==2.23.0

